Moved agents to own file, making simplified grid more like an example.
